# Reduce allocation in StringBuilder marshaling