When marshaling back results for a StringBuilder argument in a P/Invoke:
- for UTF8 it's allocating a new char[], pinning it, and then handing that off to StringBuilder.ReplaceBufferInternal, which itself then allocates a new char[], and all of that happens even if the StringBuilder's char[] is already big enough to handle the results, which is commonly the case due to the StringBuilder having been sized appropriately from the get-go.
- for both Unicode and Ansi, it's similarly allocating a new char[] invariably, even if the existing buffer is already sufficient.

This commit cleans that up.

While ideally we wouldn't use StringBuilders at all in coreclr/corefx for marshaling, we still do in some places, e.g. Dns.GetHostName.  While we should separately fix that to avoid using a StringBuilder at all, it's useful to demonstrate the impact of the change here:
```C#
[Benchmark]
public static string GetHostName() => Dns.GetHostName();
```
on my machine results in:
```
      Method |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
------------ |---------:|---------:|---------:|-------:|----------:|
 Before      | 145.1 ns | 2.923 ns | 3.801 ns | 0.2668 |   1.09 KB |
 After       | 103.0 ns | 2.142 ns | 2.859 ns | 0.1392 |     584 B |
```

cc: @jkotas, @JeremyKuhne, @vancem