@mattwarren 2018-05-09
> While ideally we wouldn't use StringBuilders at all in coreclr/corefx for marshaling

@stephentoub Just for my own interest, what's the issue with using `StringBuilder` for marshalling? Is it too heavy-weight, does it have perf concerns, is it wrong to tie-in the marshalling implementation to it or something else?