@stephentoub 2018-05-09
>  @mattwar: Just for my own interest, what's the issue with using StringBuilder for marshalling? Is it too heavy-weight, does it have perf concerns, is it wrong to tie-in the marshalling implementation to it or something else?

There are inefficiencies like the ones this PR is fixing.  But even once those are addressed, using StringBuilder in marshaling often leads to inefficiencies in app-level code, just because of the patterns it encourages.  If you already happen to have your data in a StringBuilder and you're just looking to [In] marshal into a P/Invoke, that's fine.  But often, for whatever reason, StringBuilder is used to [Out] marshal the results of some P/Invoke, and the caller often then needs to allocate a StringBuilder just for that privilege, and that in turn incurs other allocations like the char[] internally.  You can see a bunch of examples of this, e.g.

- https://github.com/dotnet/corefx/blob/bffef76f6af208e2042a2f27bc081ee908bb390b/src/System.Security.Cryptography.X509Certificates/src/Internal/Cryptography/Pal.Windows/X509Pal.X500DistinguishedName.cs#L88-L103
- https://github.com/dotnet/corefx/blob/59aeab389ae67736fe637ce275043e42506a041c/src/System.Security.Cryptography.Csp/src/System/Security/Cryptography/CapiHelper.Windows.cs#L94-L102
- https://github.com/dotnet/corefx/blob/bffef76f6af208e2042a2f27bc081ee908bb390b/src/System.Security.Cryptography.Pkcs/src/Interop/Windows/Crypt32/Interop.CertNameToStr.cs#L24-L29
- https://github.com/dotnet/corefx/blob/bffef76f6af208e2042a2f27bc081ee908bb390b/src/System.Security.Cryptography.X509Certificates/src/Internal/Cryptography/Pal.Windows/CertificatePal.cs#L296-L300

etc. You often then see devs adding more code to workaround that, like using a cache of StringBuilders, which itself incurs expense, e.g.

- https://referencesource.microsoft.com/#mscorlib/microsoft/win32/win32native.cs,746
- https://referencesource.microsoft.com/#mscorlib/system/environment.cs,631

Using a StringBuilder in the signature also prevents optimizations like first trying with stackalloc'd memory and only then going to the heap when more is needed, whereas you can do that if you just used a simple `byte*` or `char*` in the signature, and often you don't even need to go to the heap, as in the fix here, where rather than allocating a StringBuilder and allocating its internal char[] (and prior to this PR allocating another char[] when marshaling out) and allocating the resulting string, we instead just stackalloc the needed 256 bytes and then the only heap allocation we incur is the final string.

- https://github.com/dotnet/corefx/pull/29594

Plus, just using `byte*` instead of `StringBuilder` means all inputs/outputs are blittable and we no longer need any custom P/Invoke marshaling frames.