@stephentoub 2018-05-09
> I spent some time thinking about the removal of +1 for the null terminator and I think I convinced myself this looks right. The reason is that on the transition back from native we do not copy the null terminator (if it existed) so using the raw byte count should be fine.

Ok, thanks, @jeffschwMSFT.  That also addresses the rather strange issue @mikedn highlighted, where today because of the +1 you can end up with Capacity > MaxCapacity (the current code does the addition after validating the new length against the MaxCapacity rather than before.  So I'll leave the code as is.