@luqunl 2018-05-09
> Plus, just using byte* instead of StringBuilder means all inputs/outputs are blittable and we no longer need any custom P/Invoke marshaling frames.

this only works for the string itself doesn't need marshalling support. For ansi string, You cann't just use byte* instead of stringbuilder.