@stephentoub 2018-05-09
> For ansi string, You cann't just use byte* instead of stringbuilder.

Sure you can; you then just do the marshaling yourself with `new string((sbyte*)ptr)`.