@luqunl 2018-05-09
> Sure you can; you then just do the marshaling yourself, e.g. new string((sbyte*)ptr).
Thanks for this.  Do you know the perf difference between using stringbuilder and new string((sbyte*)ptr) after your change?