@luqunl 2018-05-09
> Plus, just using byte* instead of StringBuilder 

I think you mean sbyte*