@stephentoub 2018-05-09
> I think you mean sbyte*

I actually meant `byte*` as I wrote, but either would work.