@luqunl 2018-05-09
> I actually meant byte* as I wrote, but either would work.

There isn't a string constructor which takes byte* as argument.