@jkotas 2018-05-09
> Do you know the perf difference between using stringbuilder

@luqunl Real world example: https://github.com/dotnet/corefx/pull/29594 . Reduced GC allocation from 600 bytes to 40 bytes.

The overall trend in CoreFX has been to avoid builtin marshallers to get better performance, and just use blitable types. Some of the guidelines are written in https://github.com/dotnet/corefx/blob/master/Documentation/coding-guidelines/interop-guidelines.md#definition and https://github.com/dotnet/corefx/blob/master/Documentation/coding-guidelines/interop-pinvokes.md#strings