@stephentoub 2018-05-09
> There isn't a string constructor which takes byte* as argument.

I know, that's why my example does `new string((sbyte*)ptr)`.