@stephentoub 2018-05-09
Another example in corefx: https://github.com/dotnet/corefx/pull/29605