@luqunl 2018-05-09
@jkotas/ @stephentoub , I guess these Reduced GC allocation numbers doesn't include the PR/fix. is it ?
Before this PR, We recommend people to use stringbuilder as [Out] case, at least for ANSI one. 
After this PR,  what should we recommend to customer?