@luqunl 2018-05-09
> just using byte* instead of StringBuilder means all inputs/outputs are blittable 

if you want to use sbyte*, you need to make sure the managed side allocated enough memory. otherwise, you may have buffer-overrun.