@luqunl 2018-05-09
Same as Jeff, I am fine for "removal of +1 for the null terminator", since it is convert from native to managed.