@jkotas 2018-05-09
> `new string((sbyte*)ptr)`

The `sbyte` string constructor is outlier. It is more natural to use `byte`, and I believe that majority of other APIs do actually use `byte` in similar contexts.

> After this PR, what should we recommend to customer?

There is no one size fits all recommendation, and I do not think that this PR changes recommendations for writing interop code. StringBuilder is perfectly fine for code that is not performance critical, and the simplicity is preferred.

It is not different from other situations where we have slower, but more convenient APIs. E.g. When to use Linq vs. hand written for-loops?

> you need to make sure the managed side allocated enough memory. otherwise, you may have buffer-overrun.

This problem does not go away with StringBuilder. You need to ensure that when using StringBuilder as well by setting Capacity to the right amount.