@jkotas 2018-05-09
I have looked at the history where the extra null came from.

The extra null was introduced by interop refactoring in 2007. It was noticed and fixed in 2008:

```
//depot/devdiv/pu/CLR/ndp/clr/src/VM/object.cpp
change 3240200 edit on 2008/10/04 05:51:01
Title: Fix for Dev10 #520387, #519429, #519314
...
Modify ReplaceBuffer and ReplaceBufferAnsi to not allocate for the terminating null character
...
```

However, the fix was (accidentally?) reverted in:

```
//depot/devdiv/pu/CLR/ndp/clr/src/VM/object.cpp
change 3431533 edit on 2010/04/02 19:38:45

	Submitting Main --> Pu\clr FI
```

The detailed history for this was obliterated, so I am not able to tell what happened there. My bet is that it was just incorrectly resolved merge conflict.